update styling, and spec
