+++ libs/data-access/admin-api/src/index.ts
@@ -1,4 +1,4 @@
 export * from './lib/api/client';
 export * from './lib/api/hooks';
-export type * from './lib/api/type';
+export * from './lib/api/type';
 export * from './lib/AdminBaseUrlProvider';
