+++ libs/data-access/admin-api/src/lib/api/index.d.ts
@@ -1130,7 +1130,8 @@ export interface components {
       /** Format: date-time */
       next_retry_at?: string;
       id: string;
-      invoked_by: string;
+      /** @enum {string} */
+      invoked_by: 'ingress' | 'service';
       /** @enum {string} */
       status:
         | 'succeeded'
@@ -1145,8 +1146,7 @@ export interface components {
         | 'ready';
       target: string;
       target_handler_name: string;
-      /** @enum {string} */
-      target_service_key?: 'ingress' | 'service';
+      target_service_key?: string;
       target_service_name: string;
       /** @enum {string} */
       target_service_ty: 'service' | 'virtual_object' | 'workflow';
@@ -1190,7 +1190,8 @@ export interface components {
       /** Format: date-time */
       next_retry_at?: string;
       id: string;
-      invoked_by: string;
+      /** @enum {string} */
+      invoked_by: 'ingress' | 'service';
       /** @enum {string} */
       status:
         | 'pending'
@@ -1202,8 +1203,7 @@ export interface components {
         | 'completed';
       target: string;
       target_handler_name: string;
-      /** @enum {string} */
-      target_service_key?: 'ingress' | 'service';
+      target_service_key?: string;
       target_service_name: string;
       /** @enum {string} */
       target_service_ty: 'service' | 'virtual_object' | 'workflow';
