+++ libs/data-access/admin-api/src/lib/api/query.json
@@ -1617,7 +1617,8 @@
             "type": "string"
           },
           "invoked_by": {
-            "type": "string"
+            "type": "string",
+            "enum": ["ingress", "service"]
           },
           "status": {
             "type": "string",
@@ -1641,8 +1642,7 @@
             "type": "string"
           },
           "target_service_key": {
-            "type": "string",
-            "enum": ["ingress", "service"]
+            "type": "string"
           },
           "target_service_name": {
             "type": "string"
@@ -1759,7 +1759,8 @@
             "type": "string"
           },
           "invoked_by": {
-            "type": "string"
+            "type": "string",
+            "enum": ["ingress", "service"]
           },
           "status": {
             "type": "string",
@@ -1780,8 +1781,7 @@
             "type": "string"
           },
           "target_service_key": {
-            "type": "string",
-            "enum": ["ingress", "service"]
+            "type": "string"
           },
           "target_service_name": {
             "type": "string"
