+++ libs/data-access/admin-api/src/lib/api/type.ts
@@ -75,3 +75,27 @@ export type FilterNumberItem = components['schemas']['FilterNumberItem'];
 export type FilterStringListItem =
   components['schemas']['FilterStringListItem'];
 export type FilterStringItem = components['schemas']['FilterStringItem'];
+
+export type HTTPDeployment = Exclude<Deployment, { arn: string }>;
+export type LambdaDeployment = Exclude<Deployment, { uri: string }>;
+export type DeploymentType = 'uri' | 'arn';
+export function isHttpDeployment(
+  deployment: Deployment
+): deployment is HTTPDeployment {
+  return 'uri' in deployment;
+}
+export function isLambdaDeployment(
+  deployment: Deployment
+): deployment is LambdaDeployment {
+  return 'arn' in deployment;
+}
+export function getEndpoint(deployment?: Deployment) {
+  if (!deployment) {
+    return undefined;
+  }
+  if (isHttpDeployment(deployment)) {
+    return deployment.uri;
+  } else {
+    return deployment.arn;
+  }
+}
